\name{backuputils-package}
\alias{backuputils}
\alias{backuputils-package}
\title{backuputils: Backup Utilities for AWS CLI Workflows}
\description{
Wraps selected AWS CLI commands for copying local data to S3 and checking that
uploaded content matches the local source.
}
\seealso{
\code{\link{aws_sso_login}}, \code{\link{aws_copy}}, \code{\link{aws_verify_copy}}, \code{\link{backup_fastqs}}, \code{\link{depot_s3_uri}}
}
\keyword{package}
